+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 ## Changelog
 
+### v1.9 (unreleased)
+#### Added
+- option to make white noise identical across dots
+#### Changed
+- dots disabled when chair UDP server is not running or chair velocity is NaN
+
+---
+
 ### v1.8 (2024-05-15)
 #### Fixed
 - better repositioning of dots that exit the arena, to prevent bar of dots at bottom of field of view
 
----